feat(projects): Add categories for project entries (#13)
* chore: remove unsused favicon and unsused icons.
chore: Remove unused icons and files

* chore: remove unnecessary comments

* refactor: use absolute import statements

* refactor: Categorize projects on Projects page

* docs: Update technologies for one project entry

* style(projects): Update CSS for technology items

Refine the appearance of the .projects-card__tech-item element to improve
readability and visual consistency within the project cards.

* docs: Update projects data and structure

* docs: refactor to accept arrays or strings

* refactor: update rendering for descriptions.

* docs: Update image for porfolio project

* refactor: render description as array of strings

* style: update font size to improve readability

* docs: update technologies list for portfolio

* feat(projects): Add entry for old portfolio site

Add a new project entry for the previous version of the portfolio website.

* style(projects): Refine project card button styles

Improve the visual appearance and consistency of buttons within the project cards.

* docs(projects): Reorder tutorial projects by date

Update the order of projects within the 'Tutorial-based' category to display
the newest projects at the top.